LayoutTests/fast/media/mq-color-index-02-expected.txt


PASS CSSOM Sanity check
PASS (color-index: 0.0) is invalid
PASS (color-index: -1) is invalid
PASS (color-index: 2.1) is invalid
